+++ src/constants/index.ts
@@ -1,3 +1,1 @@
-export * from './authPage';
 export * from './exhibition';
-export * from './mypage';
