+++ src/types/index.ts
@@ -1,1 +1,2 @@
+export * from './exhibition';
 export * from './mypage';
